Basement.Bits

Description

Documentation

class BitOps bits where

operation over bits

Minimal complete definition

(.&.), (.|.), (.^.), (.<<.), (.>>.)

Methods

(.&.) :: bits -> bits -> bits infixl 7

(.|.) :: bits -> bits -> bits infixl 5

(.^.) :: bits -> bits -> bits infixl 6

(.<<.) :: bits -> CountOf Bool -> bits infixl 8

(.>>.) :: bits -> CountOf Bool -> bits infixl 8

bit :: Offset Bool -> bits

construct a bit set with the bit at the given index set.

default bit :: Integral bits => Offset Bool -> bits

isBitSet :: bits -> Offset Bool -> Bool

test the bit at the given index is set

default isBitSet :: (Integral bits, Eq bits) => bits -> Offset Bool -> Bool

setBit :: bits -> Offset Bool -> bits

set the bit at the given index

default setBit :: Integral bits => bits -> Offset Bool -> bits

clearBit :: bits -> Offset Bool -> bits

clear the bit at the given index

default clearBit :: FiniteBitsOps bits => bits -> Offset Bool -> bits

class FiniteBitsOps bits where

operation over finite bits

Minimal complete definition

numberOfBits, rotateL, rotateR, popCount, bitFlip

Methods

numberOfBits :: bits -> CountOf Bool

get the number of bits in the given object

rotateL :: bits -> CountOf Bool -> bits infixl 8

rotate the given bit set.

rotateR :: bits -> CountOf Bool -> bits infixl 8

rotate the given bit set.

popCount :: bits -> CountOf Bool

count of number of bit set to 1 in the given bit set.

bitFlip :: bits -> bits

reverse all bits in the argument

countLeadingZeros :: bits -> CountOf Bool

count of the number of leading zeros

default countLeadingZeros :: BitOps bits => bits -> CountOf Bool

countTrailingZeros :: bits -> CountOf Bool

count of the number of trailing zeros

default countTrailingZeros :: BitOps bits => bits -> CountOf Bool

data Bits (n :: Nat)

Bool set of n bits.

toBits :: SizeValid n => Natural -> Bits n

convert the given Natural into a Bits of size n

if bits that are not within the boundaries of the 'Bits n' will be truncated.

allOne :: forall n. SizeValid n => Bits n

construct a Bits with all bits set.

this function is equivalet to maxBound