Basement.Compat.PrimTypes

Description

Documentation

type FileSize# = Word64#

File size in bytes

type Offset# = Int#

Offset in a bytearray, string, type alias

for code documentation purpose only, just a simple type alias on Int#

type CountOf# = Int#

CountOf in bytes type alias

for code documentation purpose only, just a simple type alias on Int#

type Bool# = Int#

Lowlevel Boolean

type Pinned# = Bool#

Pinning status