Basement.Compat.Primitive

Description

Documentation

bool# :: Int# -> Bool

turn an Int# into a Bool

data PinnedStatus

Flag record whether a specific byte array is pinned or not

Constructors

Pinned

Unpinned

toPinnedStatus# :: Pinned# -> PinnedStatus

compatMkWeak# :: o -> b -> IO () -> State# RealWorld -> (# State# RealWorld, Weak# b #)

A mkWeak# version that keep working on 8.0

signature change in ghc-prim: * 0.4: mkWeak# :: o -> b -> c -> State# RealWorld -> ( RealWorld, Weak# b#) * 0.5 :mkWeak# :: o -> b -> (State# RealWorld -> ( RealWorld, c#)) -> State# RealWorld -> ( RealWorld, Weak# b#)

compatIsByteArrayPinned# :: ByteArray# -> Pinned#

compatIsMutableByteArrayPinned# :: MutableByteArray# s -> Pinned#

unsafeCoerce# :: forall (q :: RuntimeRep) (r :: RuntimeRep) (a :: TYPE q) (b :: TYPE r). a -> b

Highly, terribly dangerous coercion from one representation type to another. Misuse of this function can invite the garbage collector to trounce upon your data and then laugh in your face. You don't want this function. Really.

data Word

A Word is an unsigned integral type, with the same size as Int.

Constructors

W# Word#

data Word8# :: TYPE 'Word8Rep

data Word16# :: TYPE 'Word16Rep

data Word32# :: TYPE 'Word32Rep

data Int8# :: TYPE 'Int8Rep

data Int16# :: TYPE 'Int16Rep

data Int32# :: TYPE 'Int32Rep

word8ToWord16# :: Word8# -> Word16#

word8ToWord32# :: Word8# -> Word32#

word8ToWord# :: Word8# -> Word#

word16ToWord8# :: Word16# -> Word8#

word16ToWord32# :: Word16# -> Word32#

word16ToWord# :: Word16# -> Word#

word32ToWord# :: Word32# -> Word#

word32ToWord8# :: Word32# -> Word8#

word32ToWord16# :: Word32# -> Word16#

wordToWord32# :: Word# -> Word32#

wordToWord16# :: Word# -> Word16#

wordToWord8# :: Word# -> Word8#

int8ToInt16# :: Int8# -> Int16#

int8ToInt32# :: Int8# -> Int32#

int8ToInt# :: Int8# -> Int#

int16ToInt32# :: Int16# -> Int32#

int16ToInt# :: Int16# -> Int#

int32ToInt# :: Int32# -> Int#

intToInt8# :: Int# -> Int8#

intToInt16# :: Int# -> Int16#

intToInt32# :: Int# -> Int32#

word8ToInt# :: Word8# -> Int#

word8ToInt16# :: Word8# -> Int16#

word8ToInt32# :: Word8# -> Int32#

charToWord32# :: Char# -> Word32#

word8ToChar# :: Word8# -> Char#

word16ToChar# :: Word16# -> Char#

word32ToChar# :: Word32# -> Char#

wordToChar# :: Word# -> Char#

plusWord8# :: Word8# -> Word8# -> Word8#

uncheckedShiftRLWord16# :: Word16# -> Int# -> Word16#

plusWord16# :: Word16# -> Word16# -> Word16#

uncheckedShiftRLWord32# :: Word32# -> Int# -> Word32#

plusWord32# :: Word32# -> Word32# -> Word32#

plusInt8# :: Int8# -> Int8# -> Int8#

plusInt16# :: Int16# -> Int16# -> Int16#

plusInt32# :: Int32# -> Int32# -> Int32#