Basement.Endianness

Description

Set endianness tag to a given primitive. This will help for serialising data for protocols (such as the network protocols).

Documentation

class ByteSwap a

Class of types that can be byte-swapped.

e.g. Word16, Word32, Word64

Minimal complete definition

byteSwap

Big Endian

newtype BE a

Big Endian value

Constructors

BE

Fields

unBE :: a

toBE :: ByteSwap a => a -> BE a

Convert a value in cpu endianess to big endian

fromBE :: ByteSwap a => BE a -> a

Convert from a big endian value to the cpu endianness

Little Endian

newtype LE a

Little Endian value

Constructors

LE

Fields

unLE :: a

toLE :: ByteSwap a => a -> LE a

Convert a value in cpu endianess to little endian

fromLE :: ByteSwap a => LE a -> a

Convert from a little endian value to the cpu endianness

System Endianness

data Endianness

Constructors

LittleEndian

BigEndian

endianness :: Endianness

endianness of the current architecture