Basement.Numerical.Multiplicative

Documentation

class Multiplicative a where

Represent class of things that can be multiplied together

x * midentity = x
midentity * x = x

Minimal complete definition

midentity, (*)

Methods

midentity :: a

Identity element over multiplication

(*) :: a -> a -> a infixl 7

Multiplication of 2 elements that result in another element

(^) :: (IsNatural n, Enum n, IDivisible n) => a -> n -> a infixr 8

Raise to power, repeated multiplication e.g. > a ^ 2 = a * a > a ^ 10 = (a ^ 5) * (a ^ 5) .. (^) :: (IsNatural n) => a -> n -> a

class (Additive a, Multiplicative a) => IDivisible a where

Represent types that supports an euclidian division

(x ‘div‘ y) * y + (x ‘mod‘ y) == x

Minimal complete definition

div, mod | divMod

Methods

div :: a -> a -> a

mod :: a -> a -> a

divMod :: a -> a -> (a, a)

class Multiplicative a => Divisible a where

Support for division between same types

This is likely to change to represent specific mathematic divisions

Methods

(/) :: a -> a -> a infixl 7

recip :: Divisible a => a -> a