Basement.Numerical.Number

Documentation

class (Integral a, Eq a, Ord a) => IsIntegral a where

Number literals, convertible through the generic Integer type.

all number are Enum'erable, meaning that you can move to next element

Methods

toInteger :: a -> Integer

class IsIntegral a => IsNatural a where

Non Negative Number literals, convertible through the generic Natural type

Methods

toNatural :: a -> Natural