Basement.Terminal

Documentation

initialize :: IO ()

getDimensions :: IO (CountOf Char, CountOf Char)

Return the size of the current terminal

If the system is not supported or that querying the system result in an error then a default size of (80, 24) will be given back.