Basement.Types.AsciiString

Description

A AsciiString type backed by a ASCII encoded byte array and all the necessary functions to manipulate the string.

Documentation

newtype AsciiString

Opaque packed array of characters in the ASCII encoding

Constructors

AsciiString

Fields

toBytes :: UArray Char7

newtype MutableAsciiString st

Constructors

MutableAsciiString (MUArray Char7 st)

Binary conversion

fromBytesUnsafe :: UArray Word8 -> AsciiString

Convert a Byte Array representing ASCII data directly to an AsciiString without checking for ASCII validity

If the input contains invalid Char7 value (anything above 0x7f), it will trigger runtime async errors when processing data.

In doubt, use fromBytes

fromBytes :: UArray Word8 -> Maybe AsciiString

Convert a Byte Array representing ASCII checking validity.

If the byte array is not valid, then Nothing is returned