Basement.Types.Ptr

Documentation

data Addr

Constructors

Addr Addr#

addrPlus :: Addr -> Offset Word8 -> Addr

addrPlusSz :: Addr -> CountOf Word8 -> Addr

addrPlusCSz :: Addr -> CSize -> Addr

data Ptr a

A value of type Ptr a represents a pointer to an object, or an array of objects, which may be marshalled to or from Haskell values of type a.

The type a will often be an instance of class Storable which provides the marshalling operations. However this is not essential, and you can provide your own operations to access the pointer. For example you might write small foreign functions to get or set the fields of a C struct.

Constructors

Ptr Addr#

ptrPlus :: Ptr a -> Offset Word8 -> Ptr a

ptrPlusSz :: Ptr a -> CountOf Word8 -> Ptr a

ptrPlusCSz :: Ptr a -> CSize -> Ptr a

castPtr :: Ptr a -> Ptr b

The castPtr function casts a pointer from one type to another.