Basement.Compat.Primitive

Description

Documentation

bool# :: Int# -> Bool

turn an Int# into a Bool

Since GHC 7.8, boolean primitive don't return Bool but Int#.

data PinnedStatus

Flag record whether a specific byte array is pinned or not

Constructors

Pinned

Unpinned

toPinnedStatus# :: Pinned# -> PinnedStatus

compatAndI# :: Int# -> Int# -> Int#

A version friendly of andI#

compatQuotRemInt# :: Int# -> Int# -> (#Int#, Int##)

A version friendly of quotRemInt#

compatCopyAddrToByteArray# :: Addr# -> MutableByteArray# s -> Int# -> Int# -> State# s -> State# s

A version friendly fo copyAddrToByteArray#

only available from GHC 7.8

compatCopyByteArrayToAddr# :: ByteArray# -> Int# -> Addr# -> Int# -> State# s -> State# s

A version friendly fo copyByteArrayToAddr#

only available from GHC 7.8

compatMkWeak# :: o -> b -> IO () -> State# RealWorld -> (#State# RealWorld, Weak# b#)

A mkWeak# version that keep working on 8.0

signature change in ghc-prim: * 0.4: mkWeak RealWorld -> ( RealWorld, Weak) * 0.5 :mkWeak RealWorld -> ( RealWorld, c RealWorld -> ( RealWorld, Weak)

compatGetSizeofMutableByteArray# :: MutableByteArray# s -> State# s -> (#State# s, Int##)

compatShrinkMutableByteArray# :: MutableByteArray# s -> Int# -> State# s -> (#State# s, MutableByteArray# s#)

compatResizeMutableByteArray# :: MutableByteArray# s -> Int# -> State# s -> (#State# s, MutableByteArray# s#)

compatIsByteArrayPinned# :: ByteArray# -> Pinned#

compatIsMutableByteArrayPinned# :: MutableByteArray# s -> Pinned#

data Word :: *

A Word is an unsigned integral type, with the same size as Int.

Constructors

W# Word#