Basement.Exception

Description

Common part for vectors

Documentation

data OutOfBound

Exception during an operation accessing the vector out of bound

Represent the type of operation, the index accessed, and the total length of the vector.

Constructors

OutOfBound OutOfBoundOperation Int Int

data OutOfBoundOperation

The type of operation that triggers an OutOfBound exception.

OOB_Index: reading an immutable vector

OOB_Read: reading a mutable vector

OOB_Write: write a mutable vector

OOB_MemCopy: copying a vector

OOB_MemSet: initializing a mutable vector

Constructors

OOB_Read

OOB_Write

OOB_MemSet

OOB_MemCopy

OOB_Index

isOutOfBound :: Offset ty -> CountOf ty -> Bool

outOfBound :: OutOfBoundOperation -> Offset ty -> CountOf ty -> a

primOutOfBound :: PrimMonad prim => OutOfBoundOperation -> Offset ty -> CountOf ty -> prim a

data InvalidRecast

Constructors

InvalidRecast RecastSourceSize RecastDestinationSize

newtype RecastSourceSize

Constructors

RecastSourceSize Int

newtype RecastDestinationSize

Constructors

RecastDestinationSize Int

data NonEmptyCollectionIsEmpty

Exception for using NonEmpty assertion with an empty collection

Constructors

NonEmptyCollectionIsEmpty