Basement.From

Description

Flexible Type convertion

From is multi parameter type class that allow converting from a to b.

Only type that are valid to convert to another type should be From instance; otherwise TryFrom should be used.

Into (resp TryInto) allows the contrary instances to be able to specify the destination type before the source. This is practical with TypeApplication

Documentation

class From a b where

Class of things that can be converted from a to b.

In a valid instance, the source should be always representable by the destination, otherwise the instance should be using TryFrom

Minimal complete definition

from

Methods

from :: a -> b

type Into b a = From a b

class TryFrom a b where

Class of things that can mostly be converted from a to b, but with possible error cases.

Minimal complete definition

tryFrom

Methods

tryFrom :: a -> Maybe b

type TryInto b a = TryFrom a b

into :: Into b a => a -> b

Same as from but reverse the type variable so that the destination type can be specified first

e.g. converting:

from _ Word (10 :: Int)

into @Word (10 :: Int)

tryInto :: TryInto b a => a -> Maybe b

same as tryFrom but reversed