Basement.IntegralConv

Documentation

class IntegralDownsize a b where

Downsize an integral value

Minimal complete definition

integralDownsizeCheck

Methods

integralDownsize :: a -> b

integralDownsize :: a ~ b => a -> b

integralDownsizeCheck :: a -> Maybe b

class IntegralUpsize a b where

Upsize an integral value

The destination type b size need to be greater or equal than the size type of a

Minimal complete definition

integralUpsize

Methods

integralUpsize :: a -> b

intToInt64 :: Int -> Int64

int64ToInt :: Int64 -> Int

wordToWord64 :: Word -> Word64

word64ToWord32s :: Word64 -> Word32x2

data Word32x2

2 Word32s

Constructors

Word32x2 !Word32 !Word32

word64ToWord :: Word64 -> Word

wordToChar :: Word -> Char

wordToInt :: Word -> Int

charToInt :: Char -> Int