Basement.MutableBuilder

Documentation

newtype Builder collection mutCollection step state err a

Constructors

Builder

Fields

runBuilder :: State (Offset step, BuildingState collection mutCollection step (PrimState state), Maybe err) state a

data BuildingState collection mutCollection step state

The in-progress state of a building operation.

The previous buffers are in reverse order, and this contains the current buffer and the state of progress packing the elements inside.

Constructors

BuildingState

Fields

prevChunks :: [collection]

prevChunksSize :: !(CountOf step)

curChunk :: mutCollection state

chunkSize :: !(CountOf step)