Basement.Numerical.Number

Documentation

class (Enum a, Eq a, Ord a, Integral a) => IsIntegral a where

Number literals, convertible through the generic Integer type.

all number are Enum'erable, meaning that you can move to next element

Minimal complete definition

toInteger

Methods

toInteger :: a -> Integer

class (Enum a, Eq a, Ord a, Integral a, IsIntegral a) => IsNatural a where

Non Negative Number literals, convertible through the generic Natural type

Minimal complete definition

toNatural

Methods

toNatural :: a -> Natural