Basement.Types.Char7

Documentation

newtype Char7

ASCII value between 0x0 and 0x7f

Constructors

Char7

Fields

toByte :: Word8

toChar :: Char7 -> Char

Convert a Char7 to a unicode code point Char

fromCharMask :: Char -> Char7

Convert a Char to a Char7 ignoring all higher bits

fromChar :: Char -> Maybe Char7

Try to convert a Char to a Char7

If the code point is non ascii, then Nothing is returned.

fromByteMask :: Word8 -> Char7

Convert a Byte to a Char7 ignoring the higher bit

fromByte :: Word8 -> Maybe Char7

Try to convert Word8 to a Char7

If the byte got higher bit set, then Nothing is returned.

individual ASCII Characters

c7_LF :: Char7

c7_CR :: Char7

c7_minus :: Char7

c7_a :: Char7

c7_A :: Char7

c7_z :: Char7

c7_Z :: Char7

c7_0 :: Char7

c7_1 :: Char7

c7_2 :: Char7

c7_3 :: Char7

c7_4 :: Char7

c7_5 :: Char7

c7_6 :: Char7

c7_7 :: Char7

c7_8 :: Char7

c7_9 :: Char7